+++ package.json
@@ -7,6 +7,7 @@
   "scripts": {
     "build": "react-router build",
     "dev": "react-router dev",
+    "dev:prod-api": "FAGORD_RUST_API_DOMAIN=https://rust-api.fagord.no react-router dev",
     "start": "react-router-serve build/server/index.js",
     "typecheck": "react-router typegen && tsc",
     "prettier": "prettier --write \"app/**/*.{ts,tsx,css}\"",
